Refactor: DTO 폴더로 정리 (#14)
