+++ src/main/java/tavebalak/OTTify/program/controller/ProgramController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import tavebalak.OTTify.common.BaseResponse;
-import tavebalak.OTTify.program.dto.SearchResponseDtoV1;
-import tavebalak.OTTify.program.dto.TrendingResponseDto;
+import tavebalak.OTTify.program.dto.searchTrending.Response.SearchResponseDtoV1;
+import tavebalak.OTTify.program.dto.searchTrending.Response.TrendingResponseDto;
 import tavebalak.OTTify.program.service.ProgramShowAndSaveService;
 
 @RestController
