+++ src/main/java/tavebalak/OTTify/program/dto/searchTrending/Response/ProgramInfo.java
@@ -1,4 +1,4 @@
-package tavebalak.OTTify.program.dto;
+package tavebalak.OTTify.program.dto.searchTrending.Response;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
