+++ src/main/java/tavebalak/OTTify/program/dto/searchTrending/Response/SearchResponseDtoV1.java
@@ -1,6 +1,7 @@
-package tavebalak.OTTify.program.dto;
+package tavebalak.OTTify.program.dto.searchTrending.Response;
 
 import lombok.Data;
+import tavebalak.OTTify.program.dto.searchTrending.Response.ProgramTrendingWeekInfo;
 
 import java.util.ArrayList;
 import java.util.List;
