+++ src/main/java/tavebalak/OTTify/program/dto/searchTrending/Response/TrendingResponseDto.java
@@ -1,6 +1,8 @@
-package tavebalak.OTTify.program.dto;
+package tavebalak.OTTify.program.dto.searchTrending.Response;
 
 import lombok.Data;
+import tavebalak.OTTify.program.dto.searchTrending.Response.ProgramTrendingDayInfo;
+import tavebalak.OTTify.program.dto.searchTrending.Response.ProgramTrendingWeekInfo;
 
 import java.util.ArrayList;
 import java.util.List;
