+++ src/main/java/tavebalak/OTTify/program/dto/searchTrending/openApi/OpenApiSearchTrendingDto.java
@@ -1,4 +1,4 @@
-package tavebalak.OTTify.program.dto;
+package tavebalak.OTTify.program.dto.searchTrending.openApi;
 
 import lombok.Getter;
